Prepare to use a Function Library.
