+++ SharpMUSH.Implementation/Functions/ChannelFunctions.cs
@@ -1,4 +1,6 @@
 ﻿using SharpMUSH.Implementation.Definitions;
+using SharpMUSH.Library.Attributes;
+using SharpMUSH.Library.Definitions;
 using SharpMUSH.Library.ParserInterfaces;
 
 namespace SharpMUSH.Implementation.Functions;
