+++ SharpMUSH.Implementation/Functions/ConnectionFunctions.cs
@@ -2,6 +2,7 @@
 using OneOf.Types;
 using SharpMUSH.Implementation.Definitions;
 using SharpMUSH.Library;
+using SharpMUSH.Library.Attributes;
 using SharpMUSH.Library.Definitions;
 using SharpMUSH.Library.Extensions;
 using SharpMUSH.Library.ParserInterfaces;
