+++ SharpMUSH.Implementation/Functions/StringFunctions.cs
@@ -2,6 +2,7 @@
 using DotNext.Collections.Generic;
 using SharpMUSH.Implementation.Definitions;
 using SharpMUSH.Library;
+using SharpMUSH.Library.Attributes;
 using SharpMUSH.Library.Definitions;
 using SharpMUSH.Library.DiscriminatedUnions;
 using SharpMUSH.Library.Extensions;
@@ -10,7 +11,7 @@ using SharpMUSH.Library.Services;
 
 namespace SharpMUSH.Implementation.Functions;
 
-public static partial class Functions
+public partial class Functions
 {
 	private static readonly Dictionary<(string, string), Regex> SpeechPatternCache = new();
