+++ SharpMUSH.Implementation/Visitors/SharpMUSHParserVisitor.cs
@@ -2,6 +2,12 @@
 using Antlr4.Runtime.Misc;
 using Antlr4.Runtime.Tree;
 using Microsoft.Extensions.Logging;
+using OneOf.Types;
+using SharpMUSH.Implementation.Definitions;
+using SharpMUSH.Implementation.Functions;
+using SharpMUSH.Library.Attributes;
+using SharpMUSH.Library.Definitions;
+using SharpMUSH.Library.DiscriminatedUnions;
 using SharpMUSH.Library.ParserInterfaces;
 using static SharpMUSHParser;
 
@@ -65,9 +71,8 @@ public class SharpMUSHParserVisitor(ILogger logger, IMUSHCodeParser parser, MStr
 
 		/* await parser.NotifyService.Notify(parser.CurrentState.Executor!.Value, MModule.single(
 			$"#{parser.CurrentState.Caller!.Value.Number}! {new string(' ', parser.CurrentState.ParserFunctionDepth!.Value)}{context.GetText()} :")); */
-
-		var result =
-			await Functions.Functions.CallFunction(logger, functionName.ToLower(), source, parser, context, arguments, this);
+		
+		var result = await CallFunction(logger, functionName.ToLower(), source, context, arguments, this);
 
 		/* await parser.NotifyService.Notify(parser.CurrentState.Caller!.Value, MModule.single(
 			$"#{parser.CurrentState.Caller!.Value.Number}! {new string(' ', parser.CurrentState.ParserFunctionDepth!.Value)}{context.GetText()} => {result.Message}")); */
@@ -75,6 +80,177 @@ public class SharpMUSHParserVisitor(ILogger logger, IMUSHCodeParser parser, MStr
 		return result;
 	}
 
+	
+	/// <summary>
+	/// TODO: Optimization needed. We should at least grab the in-built ones at startup.
+	/// TODO: Move this to a Library Service implementation.
+	/// </summary>
+	/// <param name="logger"></param>
+	/// <param name="name">Function Name</param>
+	/// <param name="source">The source MarkupString</param>
+	/// <param name="context">Function Context for Depth</param>
+	/// <param name="args">Arguments</param>
+	/// <param name="visitor"></param>
+	/// <returns>The resulting CallState.</returns>
+	public async ValueTask<CallState> CallFunction(ILogger logger, string name, MString source,
+		FunctionContext context, EvaluationStringContext[] args, SharpMUSHParserVisitor visitor)
+	{
+		try
+		{
+			if (!parser.FunctionLibrary.TryGetValue(name, out var libraryMatch))
+			{
+				var discoveredFunction = DiscoverBuiltInFunction(name);
+
+				if (!discoveredFunction.TryPickT0(out var functionValue, out _))
+				{
+					return new CallState(string.Format(Errors.ErrorNoSuchFunction, name), context.Depth());
+				}
+
+				parser.FunctionLibrary.Add(name, (functionValue,false));
+				libraryMatch = parser.FunctionLibrary[name];
+			}
+
+			var (attribute, function) = libraryMatch.LibraryInformation;
+
+			var currentStack = parser.State;
+			var currentState = parser.CurrentState;
+			var contextDepth = context.Depth();
+			var stackDepth = currentStack.Count();
+			var recursionDepth = currentStack.Count(x => x.Function == name);
+
+			List<CallState> refinedArguments;
+
+			// TODO: Check Permissions here.
+
+			/* Validation, this should probably go into its own function! */
+			if (args.Length > attribute.MaxArgs)
+			{
+				// Better Error Needed.
+				return new CallState(Errors.ErrorArgRange, context.Depth());
+			}
+
+			if (args.Length < attribute.MinArgs)
+			{
+				return new CallState(string.Format(Errors.ErrorTooFewArguments, name, attribute.MinArgs, args.Length),
+					contextDepth);
+			}
+
+			if (((attribute.Flags & FunctionFlags.UnEvenArgsOnly) != 0) && (args.Length % 2 == 0))
+			{
+				return new CallState(string.Format(Errors.ErrorGotEvenArgs, name), contextDepth);
+			}
+
+			if (((attribute.Flags & FunctionFlags.EvenArgsOnly) != 0) && (args.Length % 2 != 0))
+			{
+				return new CallState(string.Format(Errors.ErrorGotUnEvenArgs, name), contextDepth);
+			}
+
+			// TODO: Reconsider where this is. We Push down below, after we have the refined arguments.
+			// But each RefinedArguments call will create a new call to this FunctionParser without depth info.
+			if (contextDepth > parser.Configuration.CurrentValue.Limit.CallLimit)
+			{
+				// TODO: Context Depth is not the correct value to use here.
+				return new CallState(Errors.ErrorCall, contextDepth);
+			}
+
+			if (stackDepth > parser.Configuration.CurrentValue.Limit.MaxDepth)
+			{
+				return new CallState(Errors.ErrorInvoke, stackDepth);
+			}
+
+			if (recursionDepth > parser.Configuration.CurrentValue.Limit.FunctionRecursionLimit)
+			{
+				return new CallState(Errors.ErrorRecursion, recursionDepth);
+			}
+
+			var stripAnsi = attribute.Flags.HasFlag(FunctionFlags.StripAnsi);
+
+			if (!attribute.Flags.HasFlag(FunctionFlags.NoParse))
+			{
+				refinedArguments = (await Task.WhenAll(args
+						.Select(async x => new CallState(
+							stripAnsi
+								? MModule.plainText2((await visitor.VisitChildren(x))?.Message ?? MModule.empty())
+								: (await visitor.VisitChildren(x))?.Message ?? MModule.empty(), x.Depth()))))
+					.DefaultIfEmpty(new CallState(MModule.empty(), context.Depth()))
+					.ToList();
+			}
+			else if (attribute.Flags.HasFlag(FunctionFlags.NoParse) && attribute.MaxArgs == 1)
+			{
+				return new CallState(
+					MModule.substring(context.Start.StartIndex, context.Stop.StopIndex - context.Start.StartIndex + 1, source),
+					contextDepth,
+					null,
+					async () => (await visitor.VisitChildren(context) ?? CallState.Empty with { Depth = context.Depth() })
+						.Message!);
+			}
+			else
+			{
+				refinedArguments = args.Select(x => new CallState(stripAnsi
+							? MModule.plainText2(MModule.substring(x.Start.StartIndex,
+								context.Stop?.StopIndex is null ? 0 : (x.Stop.StopIndex - x.Start.StartIndex + 1), source))
+							: MModule.substring(x.Start.StartIndex,
+								context.Stop?.StopIndex is null ? 0 : (x.Stop.StopIndex - x.Start.StartIndex + 1), source),
+						x.Depth(), null,
+						async () => stripAnsi
+							? (await visitor.VisitChildren(x))!.Message
+							: MModule.plainText2((await visitor.VisitChildren(x))!.Message)))
+					.DefaultIfEmpty(new CallState(MModule.empty(), context.Depth()))
+					.ToList();
+			}
+
+			if (attribute.Flags.HasFlag(FunctionFlags.DecimalsOnly) && refinedArguments.Any(a =>
+				    !decimal.TryParse(MModule.plainText(a?.Message ?? MModule.empty()), out _)))
+			{
+				return new CallState(attribute.MaxArgs > 1 ? Errors.ErrorNumbers : Errors.ErrorNumber);
+			}
+
+			if (attribute.Flags.HasFlag(FunctionFlags.IntegersOnly) && refinedArguments.Any(a =>
+				    !int.TryParse(MModule.plainText(a?.Message ?? MModule.empty()), out _)))
+			{
+				return new CallState(attribute.MaxArgs > 1 ? Errors.ErrorIntegers : Errors.ErrorInteger);
+			}
+
+			// TODO: Consider adding the ParserContexts as Arguments, so that Evaluation can be more optimized.
+			var newParser = parser.Push(new ParserState(
+				Registers: currentState.Registers,
+				IterationRegisters: currentState.IterationRegisters,
+				RegexRegisters: currentState.RegexRegisters,
+				CurrentEvaluation: currentState.CurrentEvaluation,
+				ParserFunctionDepth: parser.CurrentState.ParserFunctionDepth + 1,
+				Function: name,
+				Command: null,
+				Switches: [],
+				Arguments: new(refinedArguments.Select((value, i) =>
+						new KeyValuePair<string, CallState>(i.ToString(), value))
+					.ToDictionary()),
+				Executor: currentState.Executor,
+				Enactor: currentState.Enactor,
+				Caller: currentState.Caller,
+				Handle: currentState.Handle
+			));
+
+			var result = await function(newParser) with { Depth = contextDepth };
+
+			return result;
+		}
+		catch (Exception ex)
+		{
+			logger.LogError(ex, nameof(CallFunction));
+			return CallState.Empty;
+		}
+	}
+
+	private Option<(SharpFunctionAttribute, Func<IMUSHCodeParser, ValueTask<CallState>>)>
+		DiscoverBuiltInFunction(string name)
+	{
+		if (!parser.FunctionLibrary.TryGetValue(name, out var result) && result.IsSystem)
+			return new None();
+
+		return (result.LibraryInformation.Attribute, new Func<IMUSHCodeParser, ValueTask<CallState>>
+			(p => (ValueTask<CallState>)result.LibraryInformation.Function.Method.Invoke(null, [p, result.LibraryInformation.Attribute])!));
+	}
+	
 	public override async ValueTask<CallState?> VisitEvaluationString(
 		[NotNull] EvaluationStringContext context) => await VisitChildren(context) ?? new(
 		MModule.substring(context.Start.StartIndex,
