+++ SharpMUSH.Library/Definitions/FunctionFlags.cs
@@ -1,4 +1,4 @@
-﻿namespace SharpMUSH.Implementation.Definitions;
+﻿namespace SharpMUSH.Library.Definitions;
 
 [Flags]
 public enum FunctionFlags
