+++ SharpMUSH.Library/ParserInterfaces/IMUSHCodeParser.cs
@@ -3,6 +3,7 @@ using SharpMUSH.Library.Services;
 using System.Collections.Immutable;
 using Microsoft.Extensions.Options;
 using SharpMUSH.Configuration.Options;
+using SharpMUSH.Library.Attributes;
 
 namespace SharpMUSH.Library.ParserInterfaces;
 
@@ -20,6 +21,7 @@ public interface IMUSHCodeParser
 	ICommandDiscoveryService CommandDiscoveryService { get; }
 	IImmutableStack<ParserState> State { get; }
 	IMediator Mediator { get; }
+	LibraryService<string, Definitions.FunctionDefinition> FunctionLibrary {get;}
 	ValueTask<CallState?> CommandCommaArgsParse(MString text);
 	ValueTask<CallState?> CommandEqSplitArgsParse(MString text);
 	ValueTask<CallState?> CommandEqSplitParse(MString text);
