+++ SharpMUSH.Library/Services/ILibraryProvider.cs
@@ -1,0 +1,6 @@
+namespace SharpMUSH.Library.Services;
+
+public interface ILibraryProvider<T>
+{
+	LibraryService<string,T> Get();
+}
